Add Variations

+++ src/main/java/de/quantumrange/woocommerce/route/builder/impl/products/UpdateProductRequest.java
@@ -12,7 +12,7 @@ import java.time.format.DateTimeFormatter;
 
 public class UpdateProductRequest extends PostRequestBuilder<Product> {
 
-	public UpdateProductRequest(OAuthConfig config, long productID) {
+	public UpdateProductRequest(OAuthConfig config, int productID) {
 		super(config, Routes.Products.UPDATE, productID);
 	}
